react-viewer에서 관리할 필요가 없는 `contentType`을 demo viewer로 옮김
