+++ demo/src/constants/ContentConstants.js
@@ -20,3 +20,22 @@ export const AvailableViewType = makeConstants(updateObject(_AvailableViewType, 
     [_AvailableViewType.PAGE]: '페이지 보기 전용',
   },
 }), {});
+
+const _ContentType = {
+  WEB_NOVEL: 10,
+  COMIC: 20,
+  WEBTOON: 30,
+};
+
+export const ContentType = makeConstants(updateObject(_ContentType, {
+  _LIST: [
+    _ContentType.WEB_NOVEL,
+    _ContentType.COMIC,
+    _ContentType.WEBTOON,
+  ],
+  _STRING_MAP: {
+    [_ContentType.WEB_NOVEL]: '웹소설',
+    [_ContentType.COMIC]: '만화',
+    [_ContentType.WEBTOON]: '웹툰',
+  },
+}), {});
